// generated/prisma/models/CommunityRate.ts
export type CommunityRateModel = runtime.Types.Result.DefaultSelection<Prisma.$CommunityRatePayload>
// ...
export type $CommunityRatePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CommunityRate"
  objects: {
    baseCurrency: Prisma.$CurrencyPayload<ExtArgs>
    quoteCurrency: Prisma.$CurrencyPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    baseCurrencyId: string
    quoteCurrencyId: string
    rate: runtime.Decimal
    location: string | null
    submittedAt: Date
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["communityRate"]>
  composites: {}
}